Rename `gptq_marlin` -> `marlin` to match MoE impl
